backend/requirements.txt
fastapi~=0.115.6
uvicorn~=0.34.0
sqlalchemy~=2.0.37
pymysql~=1.1.1
pydantic~=2.10.5
PyJWT~=2.10.1
passlib~=1.7.4
requests~=2.32.3
starlette~=0.41.3
typing_extensions~=4.12.2
httpx~=0.28.1
twilio~=9.4.3
fastapi_mail

python-dotenv~=1.0.1
h11~=0.14.0
pip~=24.3.1
attrs~=24.3.0
idna~=3.10
multidict~=6.1.0
propcache~=0.2.1
sniffio~=1.3.1
click~=8.1.8
httpcore~=1.0.7
aiohttp~=3.11.11
aiohttp-retry~=2.8.3
yarl~=1.18.3
aiosignal~=1.3.2
frozenlist~=1.5.0
python-multipart~=0.0.20
aiohappyeyeballs~=2.4.4
certifi~=2024.12.14
email_validator~=2.2.0
pydantic_core~=2.27.2
anyio~=4.8.0
urllib3~=2.3.0
greenlet~=3.1.1
annotated-types~=0.7.0
charset-normalizer~=3.4.1
Jinja2~=3.1.5